requirements.txt
PySide6==6.5.2
SQLAlchemy==2.0.41
reportlab==4.4.1
qrcode==8.2
python-telegram-bot==13.15
pillow==11.1.0
pdfkit==1.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
APScheduler==3.6.3
openpyxl==3.1.2
pytest==8.2.1
black==24.4.2
isort==5.13.2
flake8==7.0.0
pre-commit==3.7.1
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0 